+++ impala/hiveserver2.py
@@ -916,7 +916,7 @@ def connect(host, port, timeout=None, use_ssl=False, ca_cert=None,
             user=None, password=None, kerberos_service_name='impala',
             auth_mechanism=None, krb_host=None, use_http_transport=False,
             http_path='', http_cookie_names=None, retries=3, jwt=None,
-            user_agent=None, get_user_custom_headers_func=None, verify_cert=False):
+            user_agent=None, get_user_custom_headers_func=None, verify_cert=True):
     log.debug('Connecting to HiveServer2 %s:%s with %s authentication '
               'mechanism', host, port, auth_mechanism)